code/requirements.txt
aniso8601==8.0.0
astroid==2.3.3
autopep8==1.5
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pylint-flask-sqlalchemy==0.2.0
python-dotenv==0.11.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
